+++ src/Components/Event.js
@@ -1,3 +1,6 @@
-export default function Event() {
+import Attendees from "./Attendees";
+
+function Event() {
   return;
 }
+export default  Event
